api completa

+++ Projeto/backend/controllers/ArtigosController.php
@@ -2,16 +2,20 @@
 
 namespace backend\controllers;
 
-use common\models\Artigo;
+use common\models\Artigos;
 use common\models\Categoria;
+use common\models\Ivas;
+use Yii;
 use yii\data\ActiveDataProvider;
 use yii\helpers\ArrayHelper;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
+use yii\web\UploadedFile;
+use app\components\MqttComponents;
 
 /**
- * ArtigoController implements the CRUD actions for Artigo model.
+ * ArtigosController implements the CRUD actions for Artigos model.
  */
 class ArtigosController extends Controller
 {
@@ -34,14 +38,15 @@ class ArtigosController extends Controller
     }
 
     /**
-     * Lists all Artigo models.
+     * Lists all Artigos models.
      *
      * @return string
      */
     public function actionIndex()
     {
+        if (Yii::$app->user->can('verArtigos')) {
         $dataProvider = new ActiveDataProvider([
-            'query' => Artigo::find(),
+            'query' => Artigos::find(),
             /*
             'pagination' => [
                 'pageSize' => 50
@@ -57,10 +62,14 @@ class ArtigosController extends Controller
         return $this->render('index', [
             'dataProvider' => $dataProvider,
         ]);
+        } else {
+
+            $this->redirect(['site/error']);
+        }
     }
 
     /**
-     * Displays a single Artigo model.
+     * Displays a single Artigos model.
      * @param int $id ID
      * @return string
      * @throws NotFoundHttpException if the model cannot be found
@@ -73,29 +82,52 @@ class ArtigosController extends Controller
     }
 
     /**
-     * Creates a new Artigo model.
+     * Creates a new Artigos model.
      * If creation is successful, the browser will be redirected to the 'view' page.
      * @return string|\yii\web\Response
      */
     public function actionCreate()
     {
-        $model = new Artigo();
+        if (Yii::$app->user->can('criarArtigos')) {
+            $model = new Artigos();
 
-        // Obtém todas as categorias
-        $categoria = Categoria::find()->all();
+            // Obtém todas as categorias
+            $categoria = Categoria::find()->all();
+            $iva = Ivas::find()->all();
 
-        if ($model->load(\Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+            if ($model->load(\Yii::$app->request->post())) {
+                $model->save();
+
+                $imagem = UploadedFile::getInstance($model , 'imagem');
+                $imagem->saveAs('C:\wamp64\www\ProjetoPSI_\frontend\web\images\materiais\\' . $imagem->name);
+                $model->imagem = $imagem->name;
+
+                $model->save();
+
+                $mqtt = new \PhpMqtt\Client\MqttClient('127.0.0.1', '1883', 'backend');
+
+                $mqtt->connect();
+                $mqtt->publish('Artigos', 'Artigo Criado'  .$model->descricao , 1);
+                $mqtt->disconnect();
+//mosquitto_sub -h 127.0.0.1 -t Artigos
+                return $this->redirect(['view', 'id' => $model->id]);
+            }
+
+
+            return $this->render('create', [
+                'model' => $model,
+                'categoria' => $categoria,
+                'iva' => $iva,
+            ]);
+        } else {
+
+            $this->redirect(['site/error']);
         }
 
-        return $this->render('create', [
-            'model' => $model,
-            'categoria' => $categoria,
-        ]);
     }
 
     /**
-     * Updates an existing Artigo model.
+     * Updates an existing Artigos model.
      * If update is successful, the browser will be redirected to the 'view' page.
      * @param int $id ID
      * @return string|\yii\web\Response
@@ -103,19 +135,33 @@ class ArtigosController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
+        if (Yii::$app->user->can('editarArtigos')) {
+            $model = $this->findModel($id);
 
-        if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+            if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->id]);
+            }
+
+
+            $mqtt = new \PhpMqtt\Client\MqttClient('127.0.0.1', '1883', 'backend');
+            $mqtt->connect();
+            $mqtt->publish('Artigos', 'Artigo Atualizado'  .$model->descricao , 1);
+            $mqtt->disconnect();
+
+            return $this->render('update', [
+                'model' => $model,
+            ]);
+
+
+
+            } else {
+
+                $this->redirect(['site/error']);
         }
-
-        return $this->render('update', [
-            'model' => $model,
-        ]);
     }
 
     /**
-     * Deletes an existing Artigo model.
+     * Deletes an existing Artigos model.
      * If deletion is successful, the browser will be redirected to the 'index' page.
      * @param int $id ID
      * @return \yii\web\Response
@@ -123,21 +169,33 @@ class ArtigosController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        if (Yii::$app->user->can('eliminarArtigos')) {
+
+         $this->findModel($id)->delete();
+
+            $mqtt = new \PhpMqtt\Client\MqttClient('127.0.0.1', '1883', 'backend');
+            $mqtt->connect();
+            $mqtt->publish('Artigos', 'Artigo Eliminado'.$this->descricao , 1);
+            $mqtt->disconnect();
+
 
         return $this->redirect(['index']);
+        } else {
+
+            $this->redirect(['site/error']);
+        }
     }
 
     /**
-     * Finds the Artigo model based on its primary key value.
+     * Finds the Artigos model based on its primary key value.
      * If the model is not found, a 404 HTTP exception will be thrown.
      * @param int $id ID
-     * @return Artigo the loaded model
+     * @return Artigos the loaded model
      * @throws NotFoundHttpException if the model cannot be found
      */
     protected function findModel($id)
     {
-        if (($model = Artigo::findOne(['id' => $id])) !== null) {
+        if (($model = Artigos::findOne(['id' => $id])) !== null) {
             return $model;
         }
